grt: clang-tidy

+++ src/grt/src/GlobalRouter.cpp
@@ -1790,10 +1790,10 @@ void GlobalRouter::updateResources(const int& init_x,
                                    odb::dbNet* db_net)
 {
   // transform from real position to grid pos of fastrouter
-  int x0 = (int) ((init_x - grid_->getXMin()) / grid_->getTileSize());
-  int y0 = (int) ((init_y - grid_->getYMin()) / grid_->getTileSize());
-  int x1 = (int) ((final_x - grid_->getXMin()) / grid_->getTileSize());
-  int y1 = (int) ((final_y - grid_->getYMin()) / grid_->getTileSize());
+  int x0 = ((init_x - grid_->getXMin()) / grid_->getTileSize());
+  int y0 = ((init_y - grid_->getYMin()) / grid_->getTileSize());
+  int x1 = ((final_x - grid_->getXMin()) / grid_->getTileSize());
+  int y1 = ((final_y - grid_->getYMin()) / grid_->getTileSize());
 
   // The last gcell is oversized and includes space that the above
   // calculation doesn't represent so correct it:
